feat(tests): add generation profile fixture utilities and related tests
